fix import and move uniir to optional
